# Improving speed by adding `string length match` before `regular expression match`