--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -9,7 +9,8 @@
 **Enhancements:**
 
 - Added the executed command line to the HTML report. (#877) (@hitenkoku)
-- Improved speed by EID matching process changed. (#882) (@fukusuket)
+- Approximately 3% speed increase by performing exact string matching on Event IDs. (#882) (@fukusuket)
+- Approximately 14% speed increase by filtering before regex usage. (#883) (@fukusuket)
 
 **Bug Fixes:**