--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "hayabusa"
-version = "2.1.0"
+version = "2.2.0-dev"
 authors = ["Yamato Security @SecurityYamato"]
 edition = "2021"
 rust-version = "1.66.0"