@hitenkoku 2023-01-17
@fukusuket thank you for your PR.
I have some suggestions. I will review so. please wait